move Block class in `models` dir
